build.py : Support --resume in container builds
Improves docker management, and re-uses build images based on the
hash of build.py, rather than making a new one each time.

Updates the build container name to be based on the source version
so that we can re-use stopped containers and resume an error'd build
without needing to start over.